backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
python-decouple==3.8
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
drf-spectacular==0.27.0
bcrypt==4.1.2
Pillow==10.2.0
whitenoise==6.6.0
gunicorn==21.2.0
